requirements.txt
fastapi
uvicorn[standard]
aiohttp
opencv-python-headless
numpy
Pillow
rembg
onnxruntime
pydantic-settings